function/requirements.txt
azure-functions
azure-storage-blob
azure-eventhub
azure-identity